feat: C.6 first strangler module = native walkability bake
nav_bake.h/.cpp ports bakeWalkability in doubles (TS float64
semantics); NativeScene::bakeNavGrid computes it at upload
(logged, no behavior change); CMake ships it in libheretek;
navbake_cli.mjs + nav_bake_probe + parity test prove identical
blocked grids across randomized scenes.

Parity caught two real precision bugs: float32 erosion margins
(0.4f != 0.4) and float-parsed scene values flipping boundary
cells — fixed by doubles end-to-end (records, obstacles, probe
argv). TS stays source of truth. ADR-1790394000001.
Native host checks pass; parity 3/3.

+++ templates/vulkan-container/app/src/main/cpp/culling.cpp
@@ -82,8 +82,13 @@ Mat4 multiply(const Mat4& a, const Mat4& b) {
 
 AABB meshAABB(const MeshRecord& mesh) {
   AABB box;
-  box.min = {mesh.px - mesh.sx * 0.5f, mesh.py - mesh.sy * 0.5f, mesh.pz - mesh.sz * 0.5f};
-  box.max = {mesh.px + mesh.sx * 0.5f, mesh.py + mesh.sy * 0.5f, mesh.pz + mesh.sz * 0.5f};
+  // Narrowing to float is intended here (GPU culling volumes).
+  box.min = {static_cast<float>(mesh.px - mesh.sx * 0.5),
+             static_cast<float>(mesh.py - mesh.sy * 0.5),
+             static_cast<float>(mesh.pz - mesh.sz * 0.5)};
+  box.max = {static_cast<float>(mesh.px + mesh.sx * 0.5),
+             static_cast<float>(mesh.py + mesh.sy * 0.5),
+             static_cast<float>(mesh.pz + mesh.sz * 0.5)};
   return box;
 }
